requirements.txt
discord.py
asyncio
requests
bs4